+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 **Author:** [James Cherti](https://www.jamescherti.com/)
 
+## WIP
+
+- Change all defvar options to defcustom
+
 ## 1.2.1
 
 * Update the auto-save timer dynamically when `easysession-save-interval` changes by centralizing the timer logic into `easysession--update-timer`. This ensures the timer is correctly cancelled and restarted based on the current mode state. A `:set` function was added to `easysession-save-interval` so that modifying the variable immediately updates the running timer without requiring a restart of the mode.